Overhaul bucket calculations
Revert earlier changes to bucket and budget calculations. Projection for
the current period only was not a useful number. There's a grey area in
the current period within which it's not possible to state categorically
that a budget is right or wrong. The earlier calculation was better for
this purpose, but admittedly did not include a useful variance to help
correct the budget.

The new Variance amount is:

- Positive (highlighted green) if the current budget balance is ever
  greater than the periodic Amount.

- Negative (highlighted red) if the budget is projected to be negative
  at the end of the period, even without any further spending.

- Blank if the balance falls between these two limits.

Variance is not summed anywhere - if the user wishes to know how much
slush they have 'spare' they should correct their budgets until none
show a variance, and then contrast the Budget total with their bank
balance.

Also corrected all test data:

- Transaction amounts all positive (should have been negative)
- Bucket references shared between barrels (should never)
- Added ids as comments to help cross-reference
- Bucket dates should always be in the future (endOf, not startOf)
- Transaction dates should always be in the past
- Simplified transaction date offsetting
- Added data to demonstrate bucket variance and highlighting